+++ src/main/java/pl/plajer/buildbattle/handlers/ChatManager.java
@@ -43,6 +43,7 @@ public class ChatManager {
 
   private static String formatPlaceholders(String message, BaseArena arena) {
     String returnString = message;
+    returnString = StringUtils.replace(returnString, "%ARENANAME%", arena.getMapName());
     returnString = StringUtils.replace(returnString, "%TIME%", Integer.toString(arena.getTimer()));
     returnString = StringUtils.replace(returnString, "%FORMATTEDTIME%", StringFormatUtils.formatIntoMMSS((arena.getTimer())));
     returnString = StringUtils.replace(returnString, "%PLAYERSIZE%", Integer.toString(arena.getPlayers().size()));
@@ -65,19 +66,7 @@ public class ChatManager {
   }
 
   public String colorMessage(String message) {
-    try {
       return ChatColor.translateAlternateColorCodes('&', LanguageManager.getLanguageMessage(message));
-    } catch (NullPointerException e1) {
-      Bukkit.getConsoleSender().sendMessage("Game message not found!");
-      if (LanguageManager.isDefaultLanguageUsed()) {
-        Bukkit.getConsoleSender().sendMessage("Please regenerate your language.yml file! If error still occurs report it to the developer!");
-      } else {
-        Bukkit.getConsoleSender().sendMessage("Locale message string not found! Please contact developer!");
-        new ReportedException(JavaPlugin.getPlugin(Main.class), e1);
-      }
-      Bukkit.getConsoleSender().sendMessage("Access string: " + message);
-      return "ERR_MESSAGE_NOT_FOUND";
-    }
   }
 
   public String colorRawMessage(String msg) {