+++ src/main/java/pl/plajer/buildbattle/menus/options/registry/biomes/BiomesRegistry.java
@@ -75,11 +75,12 @@ public class BiomesRegistry {
   }
 
   private void registerInventory() {
-    Inventory inv = Bukkit.createInventory(null, Utils.serializeInt(biomes.size()),
+    Inventory inv = Bukkit.createInventory(null, Utils.serializeInt(biomes.size() + 1),
         plugin.getChatManager().colorMessage("Menus.Option-Menu.Items.Biome.Inventory-Name"));
     for (BiomeItem item : biomes) {
       inv.addItem(item.getItemStack());
     }
+    inv.addItem(Utils.getGoBackItem());
     inventory = inv;
   }